refactor 클래스 외부로 빼는 작업 setter 메소드 제거 convention fix
